Put a "conditional" section in the basic usage.
